+++ CHANGELOG.md
@@ -1,13 +1,15 @@
 # Changelog
 
-## 2.1.0-dev
+## 2.1.0
 
 * Enhancements
+  * Added legacy option to Comeonin.Bcrypt.gen_salt so that hashes with the older $2a$ prefix can be more easily generated.
   * To try to solve the load errors after upgrading Erlang / Elixir:
     * Force C code to be recompiled every time `deps.compile` is called.
     * Added basic upgrade function to the NIF library.
   * To be compatible with the nerves project:
     * Added CROSSCOMPILE option to the Makefile.
+    * Stopped the library autoloading on compilation (requires Elixir 1.2.2).
 
 ## 2.0.0 (2015-12-17)